+++ tezosj_android/src/main/java/milfont/com/tezosj_android/data/PocketProvider.java
@@ -1,0 +1,12 @@
+package milfont.com.tezosj_android.data;
+
+import org.jetbrains.annotations.NotNull;
+
+import network.pocket.core.Pocket;
+
+class PocketProvider extends Pocket {
+    public PocketProvider(@NotNull String devID, @NotNull String network,
+                          @NotNull String[] netIds, int maxNodes, int requestTimeOut) {
+        super(devID, network, netIds, maxNodes, requestTimeOut);
+    }
+}
