+++ tezosj_android/src/main/java/milfont/com/tezosj_android/data/TezosNetwork.java
@@ -1,0 +1,5 @@
+package milfont.com.tezosj_android.data;
+
+public enum TezosNetwork {
+    MAINNET, ALPHANET
+}
